add prestige to info from socket and refactoring
