+++ src/context/reducers/player.ts
@@ -15,6 +15,7 @@ export type Player = {
   isGuildMember: boolean;
   clickBase: number;
   productionBase: number;
+  prestige: number;
 };
 
 export type SetPlayerAction = { type: "SET_PLAYER_DATA"; payload: Player };
