+++ src/context/store.ts
@@ -70,6 +70,7 @@ export const initialState: State = {
     isGuildMember: false,
     clickBase: 1,
     productionBase: 1,
+    prestige: 0,
   },
   gameMode,
   purchaseAmount: 1,
